- Reverted modification. See below.
> @@ -301,7 +301,8 @@ void UIListView::removeLastItem()
>
>  UIWidget* UIListView::getItem(unsigned int index)
>  {
> -    if ((int)index < 0 || index >= _items->count())
> +    // @todo Signed/unsigned mishmash. Index is must be unsigned int always. Yes?
> +    if ( ((int)index < 0) || ((int)index >= _items->count()) )

May be we should change the type of parameter.
So could you please revert the modification of this file?

+++ cocos/gui/UIListView.cpp
@@ -301,8 +301,7 @@ void UIListView::removeLastItem()
 
 UIWidget* UIListView::getItem(unsigned int index)
 {
-    // @todo Signed/unsigned mishmash. Index is must be unsigned int always. Yes?
-    if ( ((int)index < 0) || ((int)index >= _items->count()) )
+    if ((int)index < 0 || index >= _items->count())
     {
         return nullptr;
     }
